Changed adj matrix to uint8_t type

+++ central.cpp
@@ -74,7 +74,7 @@ using namespace std;
 	}obj[400];
 
 	struct adj_matrix{
-		int adj_m[160000];
+		uint8_t adj_m[160000];
 	}adj;
 
 	struct score_map{
